+++ affise_attribution_module_advertising/android/src/main/kotlin/com/affise/affise_attribution_module_advertising/AffiseAttributionModuleAdvertisingPlugin.kt
@@ -1,0 +1,10 @@
+package com.affise.affise_attribution_module_advertising
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+/** AffiseAttributionModuleAdvertisingPlugin */
+class AffiseAttributionModuleAdvertisingPlugin : FlutterPlugin {
+    override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) = Unit
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) = Unit
+}