+++ affise_attribution_module_advertising/ios/Classes/AffiseAttributionModuleAdvertisingPlugin.swift
@@ -1,0 +1,5 @@
+import Flutter
+
+public class AffiseAttributionModuleAdvertisingPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {}
+}