+++ affise_attribution_module_advertising/ios/affise_attribution_module_advertising/Sources/affise_attribution_module_advertising/AffiseAttributionModuleAdvertisingPlugin.swift
@@ -1,0 +1,6 @@
+import AffiseModuleAdvertising
+import Flutter
+
+public class AffiseAttributionModuleAdvertisingPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {}
+}